C2/Week4/ungraded_labs/image_classification/requirements.txt
numpy==1.16.4
Pillow==6.1.0